+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+0.1.3 (07.11.2014)
+==================
+* The class column in the input dataset automatically made into a factor.
+* Added support for generic goodness functions, to be used instead of fidelity.
+* Added support for custom predict functions, to support more advanced usage.
+
 0.1.2 (05.11.2014)
 ==================
 * Predicted class automatically added to the data frame if not present.
 * Improved check of input arguments.
 * Fixed a bug that prevented use of the svm classifier from the e1071 package with GoldenEye.
 * Updated citation information.
-
 
 0.1.0 (25.04.2014)
 ==================